requirements.txt
amqp==2.5.1
backports.csv==1.0.7
billiard==3.6.1.0
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-cors-headers==3.1.1
django-import-export==1.2.0
django-multiselectfield==0.1.9
django-rest-passwordreset==1.0.0
django-templated-mail==1.1.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
firebase-admin==3.0.0
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
grpcio==1.24.1
httplib2==0.14.0
idna==2.8
importlib-metadata==0.23
jdcal==1.4.1
kombu==4.6.5
more-itertools==7.2.0
msgpack==0.6.2
odfpy==1.4.0
openpyxl==3.0.0
Pillow==6.2.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.25.6
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
